[action] Make all decodable action fail early upon bad item type
Any action like Query or Status will parse the first byte of it payload
on start_decoding call to determine whether the payload is valid.

This make the subsequent field decoding calls much lighter as they have
one error less to handle and the validity of the item type does not have
to be checked for each decoded field.