[action] Remove all inner Action item OpCode check
The reasoning is that if you call an Action item decoding method
(Nop::start_decoding for example) it means that you are sure that the
payload you are decoding is a Nop. Thus no need to babysit the
programmer and double check his decisions.

If needed, one can easily check the opcode explicitly before calling the
decoding method.