Remove all decodable pointer based API
While I know it would be much more convenient for micro optimisations to
have these pointer based API for C lib mapping, the API is too much
maintenance, and complexifies a lot the factorization of the library.

It is thus dropped for now.
We now support only sized data decoding.